posschedstats: tweaks to which anticollision statistics values get output and autocalculated, including especially more complete and clear tracking of requests achieved
posschedule: fixed bugs in _table_matches_original_request --- start_posintTP, and put in angle wrapping for cases of like "359.99" degrees

posschedule: added in logging and tracking of cases where a positioner was moved without a request, but net zero motion. this occurs occasionally, when an uncommanded, but still working (enabled) positioner needs to automatically move out of the way of a neighbor, for anticollision (and then return to its original location)

posschedule: removed some old debugging display code. for anyone using this, I suggest much better now is in the debugger, for a given posmovetable, call its "display()" method. I put this in recently --- it gives a nice view of the table's contents