Fixed append mode for posschedstats.save(). Header row is now not repeatedly included in the append, and old rows are not re-appended.
Also, there is now a boolean flag self.clear_cache_after_save_by_append. When this flag is True (the default), data cached in posschedstats will be cleared upon saving in "append" mode. The idea is that in typical usage at the telescope, we are always appending single rows to the csv report file. In this case, we don't need to hold onto that data in memory any longer. For cases like running harness.py simulations, you want to set that flag False, so that the whole table will be kept in memory. This facilitates the summary statistics footer at the end of the simulation run.