Now *copying* the travel range list values. Took a while to find this error --- it is the root cause of positioning errors on 2020-09-21.
There was a place in posschedulestage._propose_path_adjustment() where I retrieve these limits and then directly alter that retrieved list, as part of a small, temporary calculation of how far that positioner is allowed to go if we wanted to "jog" it out of the way of its neighbor.

Unfortunately, I had overlooked the subtlety that in Python, I was just receiving a pointer to the same list object as held in the PosModel.Axis instance. So this temporary "what if" calc in the anticollision process was mangling the actual range data of the positioner!

Previously, this was not a problem --- because I had been dynamically retrieving those range limits every time. But for a speed up, a few months ago I started caching them. Thus the cache value here was susceptible to mangling. Anyway, fixed now, by simply copying the cached value whenever it is requested, so that no outside party can harm it.

A future possible safe speed-up could be to store these limits as tuples rather than lists, so that they are immutable. That would mean, however, scrubbing anywhere else in the code for cases of their direct mutation like this. Not sure if said speedup would be significant.